+++ lib/karafka/pro/routing/features/expiring/contracts/topic.rb
@@ -16,19 +16,21 @@ module Karafka
     module Routing
       module Features
         class Expiring < Base
-          # Contract to validate configuration of the expiring feature
-          class Contract < Contracts::Base
-            configure do |config|
-              config.error_messages = YAML.safe_load(
-                File.read(
-                  File.join(Karafka.gem_root, 'config', 'locales', 'pro_errors.yml')
-                )
-              ).fetch('en').fetch('validations').fetch('topic')
-            end
+          module Contracts
+            # Contract to validate configuration of the expiring feature
+            class Topic < Karafka::Contracts::Base
+              configure do |config|
+                config.error_messages = YAML.safe_load(
+                  File.read(
+                    File.join(Karafka.gem_root, 'config', 'locales', 'pro_errors.yml')
+                  )
+                ).fetch('en').fetch('validations').fetch('topic')
+              end
 
-            nested(:expiring) do
-              required(:active) { |val| [true, false].include?(val) }
-              required(:ttl) { |val| val.nil? || (val.is_a?(Integer) && val.positive?) }
+              nested(:expiring) do
+                required(:active) { |val| [true, false].include?(val) }
+                required(:ttl) { |val| val.nil? || (val.is_a?(Integer) && val.positive?) }
+              end
             end
           end
         end