+++ lib/karafka/pro/routing/features/long_running_job/contracts/topic.rb
@@ -16,18 +16,20 @@ module Karafka
     module Routing
       module Features
         class LongRunningJob < Base
-          # Rules around long-running job settings
-          class Contract < Contracts::Base
-            configure do |config|
-              config.error_messages = YAML.safe_load(
-                File.read(
-                  File.join(Karafka.gem_root, 'config', 'locales', 'pro_errors.yml')
-                )
-              ).fetch('en').fetch('validations').fetch('topic')
-            end
+          module Contracts
+            # Rules around long-running job settings
+            class Topic < Karafka::Contracts::Base
+              configure do |config|
+                config.error_messages = YAML.safe_load(
+                  File.read(
+                    File.join(Karafka.gem_root, 'config', 'locales', 'pro_errors.yml')
+                  )
+                ).fetch('en').fetch('validations').fetch('topic')
+              end
 
-            nested(:long_running_job) do
-              required(:active) { |val| [true, false].include?(val) }
+              nested(:long_running_job) do
+                required(:active) { |val| [true, false].include?(val) }
+              end
             end
           end
         end