+++ lib/karafka/routing/features/base.rb
@@ -14,9 +14,10 @@ module Karafka
           # Extends topic and builder with given feature API
           def activate
             Topic.prepend(self::Topic) if const_defined?('Topic', false)
+            ConsumerGroup.prepend(self::ConsumerGroup) if const_defined?('ConsumerGroup', false)
             Proxy.prepend(self::Builder) if const_defined?('Builder', false)
             Builder.prepend(self::Builder) if const_defined?('Builder', false)
-            Builder.prepend(Base::Expander.new(self)) if const_defined?('Contract', false)
+            Builder.prepend(Base::Expander.new(self)) if const_defined?('Contracts', false)
           end
 
           # Loads all the features and activates them