+++ lib/karafka/routing/features/base/expander.rb
@@ -30,15 +30,21 @@ module Karafka
             scope = @scope
 
             Module.new do
-              # Runs validations related to this feature on a topic
+              # Runs validations related to this feature on a routing resources
               #
               # @param block [Proc] routing defining block
               define_method :draw do |&block|
                 result = super(&block)
 
                 each do |consumer_group|
-                  consumer_group.topics.each do |topic|
-                    scope::Contract.new.validate!(topic.to_h)
+                  if scope.const_defined?('ConsumerGroup', false)
+                    scope::Contracts::ConsumerGroup.new.validate!(consumer_group.to_h)
+                  end
+
+                  if scope.const_defined?('Topic', false)
+                    consumer_group.topics.each do |topic|
+                      scope::Contracts::Topic.new.validate!(topic.to_h)
+                    end
                   end
                 end